requirements.txt
# Core Network & Scraping
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.3.0

# Data & Processing
numpy>=2.1.0
tqdm>=4.67.0

# Image & Video Processing
Pillow>=11.0.0
moviepy>=2.1.1
imageio>=2.36.0
imageio-ffmpeg>=0.5.1
decorator>=5.1.1
proglog>=0.1.10

# Support for Python 3.12+ 
setuptools>=75.0.0
python-dotenv

# Optional AI / Motion
elevenlabs
scipy
requests